third_party/WebKit/LayoutTests/http/tests/inspector/security/blank-origins-not-shown-expected.txt

Tests that blank origins aren't shown in the security panel origins list.

Group: Main Origin
Group: Unknown / Canceled
<SPAN class=title >
https://foo.test
</SPAN>
